Refactor graphical components and modifying tooltip content according to request in issue
