+++ app/components/stats/helpers/CustomTooltip.jsx
@@ -1,0 +1,59 @@
+import PropTypes from "prop-types";
+
+export default class CustomTooltip extends React.Component{
+    stylings(p){
+        return {
+            color: p.color,
+            display: 'block',
+            right: '0',
+            margin: '0',
+            padding: '0',
+        }
+    }
+
+    render() {
+        const {active} = this.props;
+        if (active) {
+            const {payload, label} = this.props;
+            const dataType = this.props.viewMode;
+
+            if(payload && label) {
+                if (dataType === 'relative') {
+                    const labelPercentage = payload.length > 1 ? 'Percentages' : 'Percentage',
+                        valueLabel = payload.length > 1 ? 'Values' : 'Value',
+                        point = payload.map((p, index) => <span style={this.stylings(p)}>Query#{index+1} {p.value ? p.value.toFixed(2) : 0}%</span>);
+
+                    return (
+                        <div className="ms__histogram-custom-tooltip">
+                            <h4>{dataType} {valueLabel}</h4>
+                            <p>Year: <span className="rightAlign">{label}</span></p>
+                            <p>{labelPercentage}: <span className="rightAlign">{point}</span></p>
+                        </div>
+                    );
+                } else {
+                    const point = payload.map((p,index) => {
+                            return (
+                                <span style={this.stylings(p)}>Query#{index+1} {p.value ? p.value : 0}</span>
+                            )
+                        }),
+                        labelTotals = payload.length > 1 ? 'Totals' : 'Total',
+                        valueLabel = payload.length > 1 ? 'Values' : 'Value';
+
+                    return (
+                        <div className="ms__histogram-custom-tooltip">
+                            <h4>{dataType} {valueLabel}</h4>
+                            <p>Year: <span className="rightAlign">{label}</span></p>
+                            <p>{labelTotals}: <span className="rightAlign">{point}</span></p>
+                        </div>
+                    );
+                }
+            }
+        }
+        return null;
+    }
+}
+CustomTooltip.propTypes = {
+    dataType: PropTypes.string,
+    payload: PropTypes.array,
+    label: PropTypes.number
+};
